Refactor NavBar component
